fn clear_and_show to state.clear_and_show to display whether filtered
