+++ src/run.rs
@@ -64,9 +64,6 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
     state.trash_dir = trash_dir_path;
     state.current_dir = arg.canonicalize()?;
 
-    //filter mode or not
-    let mut filtered = false;
-
     //Initialize num as Arc
     let nums = Num::new();
     let nums_run = Arc::new(Mutex::new(nums));
@@ -80,7 +77,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
     //Update list, print and flush
     print!("{}", cursor::Hide);
     state.update_list()?;
-    clear_and_show(&state.current_dir);
+    state.clear_and_show();
     state.list_up(nums.skip);
     state.move_cursor(&nums, BEGINNING_ROW);
     let mut init_screen = screen_run.lock().unwrap();
@@ -153,7 +150,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                     {
                         nums.go_down();
                         nums.inc_skip();
-                        clear_and_show(&state.current_dir);
+                        state.clear_and_show();
                         state.list_up(nums.skip);
                         state.move_cursor(&nums, y);
                     } else {
@@ -169,7 +166,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                     } else if y <= BEGINNING_ROW + SCROLL_POINT && nums.skip != 0 {
                         nums.go_up();
                         nums.dec_skip();
-                        clear_and_show(&state.current_dir);
+                        state.clear_and_show();
                         state.list_up(nums.skip);
                         state.move_cursor(&nums, y);
                     } else {
@@ -194,7 +191,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                 Key::Char('g') => {
                                     print!("{}", cursor::Hide);
                                     nums.reset();
-                                    clear_and_show(&state.current_dir);
+                                    state.clear_and_show();
                                     state.list_up(nums.skip);
                                     state.move_cursor(&nums, BEGINNING_ROW);
                                 }
@@ -217,7 +214,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                     if len > (state.layout.terminal_row - BEGINNING_ROW) as usize {
                         nums.skip = (len as u16) + BEGINNING_ROW - state.layout.terminal_row;
                         nums.go_bottom(len - 1);
-                        clear_and_show(&state.current_dir);
+                        state.clear_and_show();
                         state.list_up(nums.skip);
                         let cursor_pos = state.layout.terminal_row - 1;
                         state.move_cursor(&nums, cursor_pos);
@@ -240,7 +237,8 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                 }
                                 print!("{}", screen::ToAlternateScreen);
                                 print!("{}", cursor::Hide);
-                                clear_and_show(&state.current_dir);
+                                state.filtered = false;
+                                state.clear_and_show();
                                 state.update_list()?;
                                 state.list_up(nums.skip);
                                 state.move_cursor(&nums, y);
@@ -253,7 +251,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                         continue;
                                     }
                                     Ok(_) => {
-                                        let cursor_memo = if !filtered {
+                                        let cursor_memo = if !state.filtered {
                                             ParentMemo {
                                                 to_sym_dir: Some(state.current_dir.clone()),
                                                 num: *nums,
@@ -267,7 +265,6 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                             }
                                         };
                                         p_memo_v.push(cursor_memo);
-                                        filtered = false;
 
                                         state.current_dir = true_path.clone();
                                         if let Err(e) =
@@ -278,7 +275,8 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                             continue;
                                         }
                                         state.update_list()?;
-                                        clear_and_show(&state.current_dir);
+                                        state.filtered = false;
+                                        state.clear_and_show();
                                         nums.reset();
                                         state.list_up(nums.skip);
                                         state.move_cursor(&nums, BEGINNING_ROW);
@@ -293,7 +291,8 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                     }
                                     print!("{}", screen::ToAlternateScreen);
                                     print!("{}", cursor::Hide);
-                                    clear_and_show(&state.current_dir);
+                                    state.filtered = false;
+                                    state.clear_and_show();
                                     state.list_up(nums.skip);
                                     state.move_cursor(&nums, y);
                                 }
@@ -307,7 +306,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                     }
                                     Ok(_) => {
                                         //store the last cursor position and skip number
-                                        let cursor_memo = if !filtered {
+                                        let cursor_memo = if !state.filtered {
                                             ParentMemo {
                                                 to_sym_dir: None,
                                                 num: *nums,
@@ -321,7 +320,6 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                             }
                                         };
                                         p_memo_v.push(cursor_memo);
-                                        filtered = false;
 
                                         state.current_dir = item.file_path.clone();
                                         if let Err(e) =
@@ -338,19 +336,22 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                                 if state.current_dir == memo.dir_path {
                                                     nums.index = memo.num.index;
                                                     nums.skip = memo.num.skip;
-                                                    clear_and_show(&state.current_dir);
+                                                    state.filtered = false;
+                                                    state.clear_and_show();
                                                     state.list_up(nums.skip);
                                                     state.move_cursor(&nums, memo.cursor_pos);
                                                 } else {
                                                     nums.reset();
-                                                    clear_and_show(&state.current_dir);
+                                                    state.filtered = false;
+                                                    state.clear_and_show();
                                                     state.list_up(nums.skip);
                                                     state.move_cursor(&nums, BEGINNING_ROW);
                                                 }
                                             }
                                             None => {
                                                 nums.reset();
-                                                clear_and_show(&state.current_dir);
+                                                state.filtered = false;
+                                                state.clear_and_show();
                                                 state.list_up(nums.skip);
                                                 state.move_cursor(&nums, BEGINNING_ROW);
                                             }
@@ -365,21 +366,21 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                 //Go to parent directory if exists.
                 //If the list is filtered, reload current directory.
                 Key::Char('h') | Key::Left => {
-                    if filtered {
+                    if state.filtered {
                         nums.reset();
                         state.update_list()?;
-                        clear_and_show(&state.current_dir);
+                        state.filtered = false;
+                        state.clear_and_show();
                         state.list_up(nums.skip);
                         state.move_cursor(&nums, BEGINNING_ROW);
                         screen.flush()?;
-                        filtered = false;
                         continue;
                     }
                     let pre = state.current_dir.clone();
 
                     match state.current_dir.parent() {
                         Some(parent_p) => {
-                            let cursor_memo = if !filtered {
+                            let cursor_memo = if !state.filtered {
                                 ChildMemo {
                                     dir_path: pre.clone(),
                                     num: *nums,
@@ -393,7 +394,6 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                 }
                             };
                             c_memo_v.push(cursor_memo);
-                            filtered = false;
 
                             match p_memo_v.pop() {
                                 Some(memo) => {
@@ -413,7 +413,8 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                     state.update_list()?;
                                     nums.index = memo.num.index;
                                     nums.skip = memo.num.skip;
-                                    clear_and_show(&state.current_dir);
+                                    state.filtered = false;
+                                    state.clear_and_show();
                                     state.list_up(nums.skip);
                                     state.move_cursor(&nums, memo.cursor_pos);
                                 }
@@ -442,12 +443,14 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                                     .into()
                                             {
                                                 nums.skip = (nums.index - 1) as u16;
-                                                clear_and_show(&state.current_dir);
+                                                state.filtered = false;
+                                                state.clear_and_show();
                                                 state.list_up(nums.skip);
                                                 state.move_cursor(&nums, BEGINNING_ROW + 1);
                                             } else {
                                                 nums.skip = 0;
-                                                clear_and_show(&state.current_dir);
+                                                state.filtered = false;
+                                                state.clear_and_show();
                                                 state.list_up(nums.skip);
                                                 state.move_cursor(
                                                     &nums,
@@ -457,7 +460,8 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                         }
                                         None => {
                                             nums.reset();
-                                            clear_and_show(&state.current_dir);
+                                            state.filtered = false;
+                                            state.clear_and_show();
                                             state.list_up(nums.skip);
                                             state.move_cursor(&nums, BEGINNING_ROW);
                                         }
@@ -479,7 +483,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                     let mut item = state.get_item_mut(nums.index)?;
                     item.selected = true;
 
-                    clear_and_show(&state.current_dir);
+                    state.clear_and_show();
                     state.list_up(nums.skip);
                     state.move_cursor(&nums, y);
                     screen.flush()?;
@@ -510,7 +514,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                             item.selected = false;
                                         }
 
-                                        clear_and_show(&state.current_dir);
+                                        state.clear_and_show();
                                         state.list_up(nums.skip);
                                         state.move_cursor(&nums, y);
                                         screen.flush()?;
@@ -525,7 +529,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                             item.selected = false;
                                         }
 
-                                        clear_and_show(&state.current_dir);
+                                        state.clear_and_show();
                                         state.list_up(nums.skip);
                                         state.move_cursor(&nums, y + 1);
                                     }
@@ -546,7 +550,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                             item.selected = true;
                                         }
 
-                                        clear_and_show(&state.current_dir);
+                                        state.clear_and_show();
                                         state.list_up(nums.skip);
                                         state.move_cursor(&nums, y);
                                     } else {
@@ -560,7 +564,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                             item.selected = true;
                                         }
 
-                                        clear_and_show(&state.current_dir);
+                                        state.clear_and_show();
                                         state.list_up(nums.skip);
                                         state.move_cursor(&nums, y - 1);
                                     }
@@ -582,7 +586,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                                     print!("{}", cursor::Hide);
                                                     nums.reset();
                                                     state.select_from_top(start_pos);
-                                                    clear_and_show(&state.current_dir);
+                                                    state.clear_and_show();
                                                     state.list_up(nums.skip);
                                                     print!(
                                                         " {}>{}",
@@ -607,14 +611,14 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                             - state.layout.terminal_row;
                                         nums.go_bottom(len - 1);
                                         state.select_to_bottom(start_pos);
-                                        clear_and_show(&state.current_dir);
+                                        state.clear_and_show();
                                         state.list_up(nums.skip);
                                         let cursor_pos = state.layout.terminal_row - 1;
                                         state.move_cursor(&nums, cursor_pos);
                                     } else {
                                         nums.go_bottom(len - 1);
                                         state.select_to_bottom(start_pos);
-                                        clear_and_show(&state.current_dir);
+                                        state.clear_and_show();
                                         state.list_up(nums.skip);
                                         state.move_cursor(&nums, len as u16 + BEGINNING_ROW - 1);
                                     }
@@ -637,7 +641,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                         break;
                                     }
 
-                                    clear_and_show(&state.current_dir);
+                                    state.clear_and_show();
                                     state.update_list()?;
                                     let new_len = state.list.len();
                                     if usize::from(nums.skip) >= new_len {
@@ -689,7 +693,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
 
                                 Key::Esc => {
                                     state.reset_selection();
-                                    clear_and_show(&state.current_dir);
+                                    state.clear_and_show();
                                     state.list_up(nums.skip);
                                     state.move_cursor(&nums, y);
                                     break;
@@ -716,7 +720,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                     }
                     state.update_list()?;
                     nums.reset();
-                    clear_and_show(&state.current_dir);
+                    state.clear_and_show();
                     state.list_up(nums.skip);
                     state.move_cursor(&nums, BEGINNING_ROW);
                 }
@@ -763,7 +767,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                             continue;
                                         }
 
-                                        clear_and_show(&state.current_dir);
+                                        state.clear_and_show();
                                         state.update_list()?;
                                         state.list_up(nums.skip);
                                         let cursor_pos = if state.list.is_empty() {
@@ -839,7 +843,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                         continue;
                     }
 
-                    clear_and_show(&state.current_dir);
+                    state.clear_and_show();
                     state.update_list()?;
                     state.list_up(nums.skip);
 
@@ -901,7 +905,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                     }));
 
                                     print!("{}", cursor::Hide);
-                                    clear_and_show(&state.current_dir);
+                                    state.clear_and_show();
                                     state.update_list()?;
                                     state.list_up(nums.skip);
                                     state.move_cursor(&nums, y);
@@ -971,6 +975,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                     }
                     print!(" {}{}/", cursor::Goto(2, 2), clear::CurrentLine,);
                     print!("{}", cursor::Show);
+                    state.filtered = true;
                     screen.flush()?;
 
                     let original_list = state.list.clone();
@@ -985,7 +990,6 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                         if let Some(Ok(key)) = input {
                             match key {
                                 Key::Char('\n') => {
-                                    filtered = true;
                                     reset_info_line();
                                     screen.flush()?;
 
@@ -995,13 +999,12 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                 }
 
                                 Key::Esc => {
-                                    clear_and_show(&state.current_dir);
+                                    state.filtered = false;
+                                    state.clear_and_show();
                                     state.list = original_list;
                                     state.list_up(nums.skip);
-
                                     print!("{}", cursor::Hide);
                                     state.move_cursor(&nums, y);
-
                                     break;
                                 }
 
@@ -1030,7 +1033,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                         .filter(|entry| entry.file_name.contains(result))
                                         .collect();
 
-                                    clear_and_show(&state.current_dir);
+                                    state.clear_and_show();
                                     state.list_up(0);
 
                                     print!(
@@ -1058,7 +1061,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                         .collect();
 
                                     nums.reset_skip();
-                                    clear_and_show(&state.current_dir);
+                                    state.clear_and_show();
                                     state.list_up(nums.skip);
 
                                     print!(
@@ -1148,7 +1151,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                             print_warning(e, y);
                                             break 'command;
                                         }
-                                        clear_and_show(&state.current_dir);
+                                        state.clear_and_show();
                                         state.list_up(nums.skip);
                                         print!("{}", cursor::Hide);
                                         state.move_cursor(&nums, BEGINNING_ROW);
@@ -1157,7 +1160,8 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                         //reload current dir
                                         nums.reset();
                                         state.update_list()?;
-                                        clear_and_show(&state.current_dir);
+                                        state.filtered = false;
+                                        state.clear_and_show();
                                         state.list_up(nums.skip);
                                         print!("{}", cursor::Hide);
                                         state.move_cursor(&nums, BEGINNING_ROW);
@@ -1224,7 +1228,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                             }
                                         }
                                         print!("{}", cursor::Hide);
-                                        clear_and_show(&state.current_dir);
+                                        state.clear_and_show();
                                         state.list_up(nums.skip);
                                         state.move_cursor(&nums, y);
                                         break 'command;
@@ -1253,7 +1257,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                         state.current_dir = dirs::home_dir().unwrap();
                                         nums.reset();
                                         state.update_list()?;
-                                        clear_and_show(&state.current_dir);
+                                        state.clear_and_show();
                                         state.list_up(nums.skip);
                                         print!("{}", cursor::Hide);
                                         state.move_cursor(&nums, BEGINNING_ROW);
@@ -1294,7 +1298,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                                         state.update_list()?;
                                                         p_memo_v = Vec::new();
                                                         c_memo_v = Vec::new();
-                                                        clear_and_show(&state.current_dir);
+                                                        state.clear_and_show();
                                                         state.list_up(nums.skip);
                                                         print!("{}", cursor::Hide);
                                                         state.move_cursor(&nums, BEGINNING_ROW);
@@ -1344,7 +1348,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                         reset_info_line();
                                         print!("{}", cursor::Hide);
                                         if state.current_dir == state.trash_dir {
-                                            clear_and_show(&state.current_dir);
+                                            state.clear_and_show();
                                             state.update_list()?;
                                             state.list_up(nums.skip);
                                             print_info("Trash dir emptied", y);
@@ -1371,7 +1375,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                     {
                                         print!("{}", screen::ToAlternateScreen);
 
-                                        clear_and_show(&state.current_dir);
+                                        state.clear_and_show();
                                         state.update_list()?;
                                         state.list_up(nums.skip);
 
@@ -1382,7 +1386,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                                     print!("{}", screen::ToAlternateScreen);
                                     info!("SHELL: {} {:?}", c, args);
 
-                                    clear_and_show(&state.current_dir);
+                                    state.clear_and_show();
                                     state.update_list()?;
                                     state.list_up(nums.skip);
 
@@ -1533,7 +1537,7 @@ pub fn run(arg: PathBuf, log: bool) -> Result<(), FxError> {
                         }
                     }
                     nums.reset();
-                    clear_and_show(&state.current_dir);
+                    state.clear_and_show();
                     state.list_up(nums.skip);
                     state.move_cursor(&nums, BEGINNING_ROW);
                 }
